backend/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
google-auth==1.27.1
requests==2.25.1
google-cloud-spanner==3.1.0